requirements.txt
langgraph
langchain
langchain-openai
python-dotenv
requests
beautifulsoup4
pandas
matplotlib
seaborn
yfinance